pyproject.toml
[project]
name = "langchain-udemy-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-openai>=0.3.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
    "streamlit-chat>=0.1.1",
]
